% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cibersort.R
\name{CIBERSORT}
\alias{CIBERSORT}
\title{Run CIBERSORT Immune Cell Deconvolution}
\usage{
CIBERSORT(x, c1, c2, c3, ref, meth, act)
}
\arguments{
\item{x}{A \code{tibble} of expression data.}

\item{c1}{Character string. Name of the sample column.}

\item{c2}{Character string. Name of the gene/transcript column.}

\item{c3}{Character string. Name of the abundance column.}

\item{ref}{A reference matrix for deconvolution.
Use \code{surVisualiser::my_ref} for the bundled reference.}

\item{meth}{Character string. Deconvolution method
(passed to \code{tidybulk::deconvolve_cellularity}).}

\item{act}{Character string. Action parameter
(passed to \code{tidybulk::deconvolve_cellularity}).}
}
\value{
A \code{tibble} with estimated cell-type proportions.
}
\description{
Aggregates transcript counts and performs immune cell deconvolution
using \code{tidybulk::deconvolve_cellularity}.
}
\examples{
\dontrun{
deconv_result <- CIBERSORT(
  expr_data,
  c1 = "sample",
  c2 = "symbol",
  c3 = "raw_count",
  ref = surVisualiser::my_ref,
  meth = "cibersort",
  act = "get"
)
}
}
