% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_ref}
\alias{my_ref}
\title{CIBERSORT Reference Matrix}
\format{
A matrix or data frame with genes in rows and reference
  cell types in columns.
}
\source{
Custom reference matrix for TCGA analyses.
}
\usage{
my_ref
}
\description{
A reference expression matrix for immune cell deconvolution using
the CIBERSORT algorithm via \code{tidybulk::deconvolve_cellularity}.
}
\examples{
data(my_ref)
dim(my_ref)
}
\keyword{datasets}
